+++ UNQShop/src/envios/RetiroEnSucursal.java
@@ -10,14 +10,12 @@ public class RetiroEnSucursal implements MetodoEnvio {
     }
 
     @Override
-    public double calcularCosto(Pedido pedido) {
-        // El costo de envío es siempre cero
+    public float calcularCosto(Pedido pedido) {
         return 0f; 
     }
 
     @Override
     public int estimarDiasEntrega(Pedido pedido) {
-        // Inmediato (0) si hay stock, hasta 3 días si no.
         if (sucursal.tieneStock(pedido)) {
             return 0;
         } else {
